refactor: extract stream to buffer to own module
