+++ server/lib/misc/stream-to-buffer.js
@@ -1,0 +1,12 @@
+module.exports = async readableStream => {
+  return new Promise((resolve, reject) => {
+    const chunks = []
+    readableStream.on('data', data => {
+      chunks.push(data instanceof Buffer ? data : Buffer.from(data))
+    })
+    readableStream.on('end', () => {
+      resolve(Buffer.concat(chunks))
+    })
+    readableStream.on('error', reject)
+  })
+}
